requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.7
djangorestframework==3.12.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.0
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
wcwidth==0.2.5
